requirements.txt
langchain==0.1.20
langchain-google-genai==1.0.1
python-dotenv==1.0.0
